# contracts/product/v1/workspace-private.openapi.yaml
openapi: 3.1.2
info:
  title: Cyrene Yield Workspace Private Product API
  version: 1.0.0
  description: >-
    Private service-to-service routes used by the Workspace Product projection.
    A configured bearer resolves to a fixed organization and Workspace on the
    Product server. The bearer authenticates the Platform service only; it
    carries no user, member, or role identity.
paths:
  /internal/workspace/v1/training-drafts:
    post:
      operationId: workspaceCreateTrainingDraft
      summary: Create a scoped Yield-owned TrainingDraft
      description: >-
        Product assigns the authenticated organization and Workspace scope;
        caller-supplied workspace fields are ignored. Resource, scope provenance,
        and idempotency receipt are committed in one ProductStore transaction.
      security:
        - WorkspaceServiceBearer: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./openapi.yaml#/components/schemas/CreateTrainingDraft
      responses:
        '201':
          description: Scoped Product-owned TrainingDraft created or replayed
          content:
            application/json:
              schema:
                $ref: ./openapi.yaml#/components/schemas/TrainingDraft
        '401':
          description: Workspace service bearer is missing or invalid
        '409':
          description: Idempotency key conflicts with a different request
        '422':
          description: Draft request is invalid
        '503':
          description: Workspace service credential is not configured
  /internal/workspace/v1/training-drafts/actions/import-llama-factory:
    post:
      operationId: workspaceImportLlamaFactory
      summary: Create a scoped draft from an admitted LLaMA Factory YAML subset
      description: Product assigns scope from the authenticated service credential.
      security:
        - WorkspaceServiceBearer: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, datasetVersion, yaml]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                datasetVersion:
                  $ref: ./openapi.yaml#/components/schemas/DatasetVersionRef
                yaml:
                  type: string
                  minLength: 1
                  maxLength: 1000000
      responses:
        '201':
          description: Scoped Product-owned TrainingDraft created or replayed
          content:
            application/json:
              schema:
                $ref: ./openapi.yaml#/components/schemas/TrainingDraft
        '400':
          description: YAML contains a field or value outside the admitted subset
        '401':
          description: Workspace service bearer is missing or invalid
        '409':
          description: Idempotency key conflicts with a different request
        '422':
          description: Draft request is invalid
        '503':
          description: Workspace service credential is not configured
  /internal/workspace/v1/training-drafts/{draft_id}:
    get:
      operationId: workspaceGetDraft
      summary: Read one scoped Yield-owned TrainingDraft
      security:
        - WorkspaceServiceBearer: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
      responses:
        '200':
          description: Current Product-owned TrainingDraft projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceTrainingDraftProjection'
        '401':
          description: Workspace service bearer is missing or invalid
        '404':
          description: TrainingDraft does not exist in the authenticated scope
        '503':
          description: Workspace service credential is not configured
    patch:
      operationId: workspacePrepareDraft
      summary: Prepare a scoped TrainingDraft
      security:
        - WorkspaceServiceBearer: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./openapi.yaml#/components/schemas/PrepareTrainingDraft
      responses:
        '200':
          description: Scoped draft updated by its Product owner
          content:
            application/json:
              schema:
                $ref: ./openapi.yaml#/components/schemas/TrainingDraft
        '401':
          description: Workspace service bearer is missing or invalid
        '404':
          description: TrainingDraft does not exist in the authenticated scope
        '422':
          description: Preparation request or referenced Artifact is invalid
        '503':
          description: Workspace service credential is not configured
  /internal/workspace/v1/training-drafts/{draft_id}/actions/start:
    post:
      operationId: workspaceStartRun
      summary: Start the Yield-owned run for a scoped TrainingDraft
      description: >-
        Preserves the existing 202 Product response. This operation does not
        declare an Idempotency-Key or a server-side idempotency guarantee.
      security:
        - WorkspaceServiceBearer: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
      responses:
        '202':
          description: Product accepted the run start request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceTrainingRunProjection'
        '401':
          description: Workspace service bearer is missing or invalid
        '404':
          description: TrainingDraft does not exist in the authenticated scope
        '422':
          description: Draft is not prepared or the Product operation is invalid
        '503':
          description: Workspace service credential or execution is unavailable
components:
  securitySchemes:
    WorkspaceServiceBearer:
      type: http
      scheme: bearer
      description: >-
        A dedicated high-entropy service credential whose SHA-256 digest is
        configured with one fixed organizationId and workspaceId on Yield.
  parameters:
    DraftId:
      name: draft_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 200
  schemas:
...
    WorkspaceDatasetVersionProjection:
      type: object
      additionalProperties: false
      required: [id, resourceVersion, artifact, format]
      properties:
        id:
          type: string
          format: uuid
        resourceVersion:
          type: integer
          minimum: 1
        artifact:
          $ref: '#/components/schemas/WorkspaceArtifactProjection'
        validationArtifact:
          $ref: '#/components/schemas/WorkspaceArtifactProjection'
        format:
          type: string
          enum: [ALPACA_JSONL]
...
    WorkspaceTrainingConfigurationProjection:
      type: object
      additionalProperties: false
      required: [baseModel, parameters]
      properties:
        baseModel:
          $ref: '#/components/schemas/WorkspaceBaseModelProjection'
        parameters:
          $ref: '#/components/schemas/WorkspaceTrainingParameters'
    WorkspaceImportedParametersProjection:
      type: object
      additionalProperties: false
      required: [parameters]
      properties:
        parameters:
          $ref: '#/components/schemas/WorkspaceTrainingParameters'
    WorkspaceTrainingDraftProjection:
      type: object
      additionalProperties: false
      required: [id, state, name, resourceVersion, datasetVersion, createdAt]
      properties:
        id:
          type: string
          format: uuid
        state:
          type: string
          enum: [DRAFT, PREPARED, STARTED]
        name:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '^[^/\\:]{1,200}$'
        resourceVersion:
          type: integer
          minimum: 1
        datasetVersion:
          $ref: '#/components/schemas/WorkspaceDatasetVersionProjection'
        configuration:
          $ref: '#/components/schemas/WorkspaceTrainingConfigurationProjection'
        importedParameters:
          $ref: '#/components/schemas/WorkspaceImportedParametersProjection'
        trainingRunId:
          type: string
          format: uuid
        cancellationRequestedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    WorkspaceTrainingSpecProjection:
      type: object
      additionalProperties: false
      required: [modelArtifact, datasetVersion, method, finetuningType, parameters]
      properties:
        modelArtifact:
          $ref: '#/components/schemas/WorkspaceArtifactProjection'
        datasetVersion:
          $ref: '#/components/schemas/WorkspaceDatasetVersionProjection'
        method:
          type: string
          enum: [SFT]
        finetuningType:
          type: string
          enum: [LORA]
        parameters:
          $ref: '#/components/schemas/WorkspaceTrainingParameterProjection'
    WorkspaceProducedArtifactProjection:
      type: object
      additionalProperties: false
      required: [kind, artifact, derivedFromDigests]
      properties:
        kind:
          type: string
          enum: [MODEL_CHECKPOINT, MODEL_ADAPTER, METRICS, TRAINING_MANIFEST]
        artifact:
          $ref: '#/components/schemas/WorkspaceArtifactProjection'
        derivedFromDigests:
          type: array
          items:
            type: string
            pattern: '^sha256:[0-9a-f]{64}$'
    WorkspaceTrainingFailureProjection:
      type: object
      additionalProperties: false
      required: [code, retryable]
      properties:
        code:
          type: string
          pattern: '^[A-Z0-9_]{1,100}$'
        retryable:
          type: boolean
    WorkspaceTrainingResultProjection:
      type: object
      additionalProperties: false
      required: [id, adapterArtifact, checkpointArtifacts, createdAt]
      properties:
        id:
          type: string
          format: uuid
        modelVersionDigest:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        adapterArtifact:
          $ref: '#/components/schemas/WorkspaceArtifactProjection'
        checkpointArtifacts:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceArtifactProjection'
        createdAt:
          type: string
          format: date-time
    WorkspaceTrainingRunProjection:
      type: object
      additionalProperties: false
      required:
        - id
        - draftId
        - state
        - spec
        - attemptCount
        - outputArtifacts
        - createdAt
        - updatedAt
        - resourceVersion
      properties:
        id:
          type: string
          format: uuid
        draftId:
          type: string
          format: uuid
        state:
          type: string
          enum: [QUEUED, RUNNING, COMPLETED, FAILED, CANCELLING, CANCELLED, AWAITING_RETRY]
        spec:
          $ref: '#/components/schemas/WorkspaceTrainingSpecProjection'
        attemptCount:
          type: integer
          minimum: 0
        outputArtifacts:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceProducedArtifactProjection'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resourceVersion:
          type: integer
          minimum: 1
        cancellationRequestedAt:
          type: string
          format: date-time
        result:
          $ref: '#/components/schemas/WorkspaceTrainingResultProjection'
        failure:
          $ref: '#/components/schemas/WorkspaceTrainingFailureProjection'